Additional work for #487
One loop was missed during #487, being lost in a git stash.

+++ src/workerd/api/form-data.c++
@@ -427,7 +427,7 @@ void FormData::forEach(
   // it up. Using the classic for (;;) syntax here allows for that. However, this does
   // mean that it's possible for a user to trigger an infinite loop here if new items
   // are added to the search params unconditionally on each iteration.
-  for (int i = 0; i < this->data.size(); i++) {
+  for (size_t i = 0; i < this->data.size(); i++) {
     auto& [key, value] = this->data[i];
     static constexpr auto ARG_COUNT = 3;
